Checkpoing restructuring to use tasting queue
Replaces all web server threads racing for a Mutex with them ... racing
for a Mutex to write to a shared channel. The reason we can't do any
better at the moment (i.e., give a separate TX end to the channel to
every web server thread) is that I can't see a way of passing a
bootstrap argument to the `hyper` threads that `afterparty` creates.